Creation of STATS matrix; Σ+ and Σ- functional; sums are functional;... Undo not yet considered; Tracking STATS matrix changes not yet done.

See merge request Over_score/wp43s!1239

+++ src/wp43s/browsers/registerBrowser.c
@@ -1,4 +1,4 @@
-/* This file is part of 43S.
+  /* This file is part of 43S.
  *
  * 43S is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -177,7 +177,7 @@
           }
           else {
             dataBlock_t* dblock = REGISTER_COMPLEX34_MATRIX_DBLOCK(regist);
-            sprintf(tmpString, "%" PRIu16 " element%s " STD_CORRESPONDS_TO " 4+%" PRIu32 " bytes", dblock->matrixRows * dblock->matrixColumns, (dblock->matrixRows * dblock->matrixColumns)==1 ? "" : "s", (uint32_t)TO_BYTES(dblock->matrixRows * dblock->matrixColumns * COMPLEX34_SIZE));
+            sprintf(tmpString, "%" PRIu16 " element%s " STD_CORRESPONDS_TO " 4+%" PRIu32 " bytes", (uint16_t)(dblock->matrixRows * dblock->matrixColumns), (dblock->matrixRows * dblock->matrixColumns)==1 ? "" : "s", (uint32_t)TO_BYTES(dblock->matrixRows * dblock->matrixColumns * COMPLEX34_SIZE));
           }
         }
         else if(getRegisterDataType(regist) == dtConfig) {
@@ -291,7 +291,7 @@
               }
               else {
                 dataBlock_t* dblock = REGISTER_COMPLEX34_MATRIX_DBLOCK(regist);
-                sprintf(tmpString, "%" PRIu16 " element%s " STD_CORRESPONDS_TO " 4+%" PRIu32 " bytes", dblock->matrixRows * dblock->matrixColumns, (dblock->matrixRows * dblock->matrixColumns)==1 ? "" : "s", (uint32_t)TO_BYTES(dblock->matrixRows * dblock->matrixColumns * COMPLEX34_SIZE));
+                sprintf(tmpString, "%" PRIu16 " element%s " STD_CORRESPONDS_TO " 4+%" PRIu32 " bytes", (uint16_t)(dblock->matrixRows * dblock->matrixColumns), (dblock->matrixRows * dblock->matrixColumns)==1 ? "" : "s", (uint32_t)TO_BYTES(dblock->matrixRows * dblock->matrixColumns * COMPLEX34_SIZE));
               }
             }
             else if(getRegisterDataType(regist) == dtConfig) {
